requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
beautifulsoup4
requests
google-generativeai
playwright
